Documents/Local dev projects/Content machine/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
numpy==1.24.3
Pillow==10.0.1
python-dateutil==2.8.2
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.6
gnews==0.2.9
schedule==1.2.0
